feat: test code for lambda get imgs list
