+++ src/lambda/sync-image/utils.py
